Refactor classes to allow for event_name to be assumed by the library instead of explicitly passed in
