+++ README.md
@@ -1,38 +1,57 @@
 # CleverEventsRails
-
-TODO: Delete this and the text below, and describe your gem
-
-Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/clever_events_rails`. To experiment with that code, run `bin/console` for an interactive prompt.
-
 ## Installation
 
-TODO: Replace `UPDATE_WITH_YOUR_GEM_NAME_IMMEDIATELY_AFTER_RELEASE_TO_RUBYGEMS_ORG` with your gem name right after releasing it to RubyGems.org. Please do not do it earlier due to security reasons. Alternatively, replace this section with instructions to install your gem from git if you don't plan to release to RubyGems.org.
-
-Install the gem and add to the application's Gemfile by executing:
-
-```bash
-bundle add UPDATE_WITH_YOUR_GEM_NAME_IMMEDIATELY_AFTER_RELEASE_TO_RUBYGEMS_ORG
+Add the gem to your gemfile:
+```ruby
+gem "clever_events_rails", "~> 0.1.0", git: "https://github.com/clever-real-estate/clever-events-rails"
 ```
 
-If bundler is not being used to manage dependencies, install the gem by executing:
-
-```bash
-gem install UPDATE_WITH_YOUR_GEM_NAME_IMMEDIATELY_AFTER_RELEASE_TO_RUBYGEMS_ORG
+```
+bundle install
 ```
 
 ## Usage
+You can include the gem's module in the model you want to publish events from:
 
-TODO: Write usage instructions here
+```ruby
+class Object < ApplicationRecord
+  include CleverEvents::Publishable
+  ...
+end
+```
+
+Simply including the `Publishable` module it will give access to a few methods:
+- `publishable_attrs` is a class attribute, that accepts an array of symbols corresponding to the attributes we want to send events about (updates only).
+- `publishable_actions` is another class attribute that accepts an array of symbols, corresponding to the _actions_ that the object undergoes. If this is not explicitly added, the defaults are `[:create, :update, :destroy]`.
+- `#publish_event` is automatically included. It synchrounously publishes an event via the adapter specified.
+
+If you want to implement your own `#publish_event` method, just implement it in the model:
+```ruby
+def publish_event!
+  SomeEventPublisherJob.perform_later(event_name, self) # or whatever implementation you want
+end
+```
+> Note: `event_name` is a method included in whatever class includes CleverEvents::Publisher, which outputs a name in the structure of `object_class.action`, ex: `TestObject.updated`.
+
+And then make sure your implementation eventually calls `.publish_event!`:
+```ruby
+class SomeEventPublisherJob
+  include CleverEvents::Publishaer
+
+  def perform_later(event_name, object)
+    CleverEvents::Publisher.publish_event!(event_name, object)
+  end
+end
 
 ## Development
 
 After checking out the repo, run `bin/setup` to install dependencies. Then, run `rake spec` to run the tests. You can also run `bin/console` for an interactive prompt that will allow you to experiment.
 
-To install this gem onto your local machine, run `bundle exec rake install`. To release a new version, update the version number in `version.rb`, and then run `bundle exec rake release`, which will create a git tag for the version, push git commits and the created tag, and push the `.gem` file to [rubygems.org](https://rubygems.org).
+To install this gem onto your local machine, run `bundle exec rake install`.
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at https://github.com/[USERNAME]/clever_events_rails.
+Bug reports and pull requests are welcome on GitHub at https://github.com/clever-real-estate/clever_events_rails.
 
 ## License
