+++ lib/clever_events_rails/clever_events/adapters.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module CleverEvents
+  module Adapters
+    require "clever_events_rails/clever_events/adapters/sns_adapter"
+  end
+end
