+++ lib/clever_events_rails/clever_events/message.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Events
+module CleverEvents
   class Message
     attr_reader :event_name, :entity
