+++ lib/clever_events_rails/clever_events/publisher.rb
@@ -1,26 +1,9 @@
 # frozen_string_literal: true
 
-module Events
+module CleverEvents
   module Publisher
     extend ActiveSupport::Concern
-    include Events::Publisher::Publishable
-    include Events::Adapters::SnsAdapter
-
-    included do
-      after_commit do
-        publish_event! if publish_event?
-      end
-
-      def publish_event!
-        Events::Publisher.publish_event!(event_name, self)
-      end
-
-      private
-
-      def event_name
-        "#{self.class.name}.#{event_type}"
-      end
-    end
+    include CleverEvents::Adapters::SnsAdapter
 
     class << self
       def publish_event!(event_name, entity)
@@ -39,7 +22,7 @@ module Events
       end
 
       def event_adapter
-        Events::Adapters::SnsAdapter
+        CleverEvents::Adapters::SnsAdapter
       end
     end
   end
