--- lib/clever_events_rails/events/adapters.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Events
-  module Adapters
-    require "clever_events_rails/events/adapters/sns_adapter"
-  end
-end
