+++ spec/dummy_app/app/models/test_object.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class TestObject < ActiveRecord::Base
-  include Events::Publisher
+  include CleverEvents::Publishable
 
   publishable_attrs :first_name, :last_name, :email, :phone
   publishable_actions :created, :updated, :destroyed
